Arm and Lifter code
Soft points work with the arm.

The lift currently does not work with set points because the string pot needs to be connected directly to the talons.

The logic for the arm's soft limits works.

lifter joystickControl3 is the one that currently works. 1 is being used for set points testing.