Lifter soft points work
The string pot attached to the Talonworks perfectly.

The logic needs to be figured out still, and I wouldlike to go over it with everyone.

Comment:

I started using the BigBrother subsytem (used for looping set points) to start making simple arrays for se and soft points, mainly for learning purposes. I'm not sure if we'll continue to use this subsytem or not.