Trolley code works
The array has been created with all of the set points. There are 21 set points that are set to the throttle.

The lift, trolley, and arm set points are all set based onthe thorttle's position.

I changed the names of the the lifter to trolley

BigBrother Subsytem has the array initalized in it

BigBrother Command has the logic fro the array.

The array consists of:
lift, arm, and trolley set points
arm and trolley soft limits
arm and trolley fine adjustment

CAN IDs should be the same as the comp bot / practice bot, but it is best to check before testing